pyproject.toml
[tool.poetry]
name = "bruno-grobelsek"
version = "0.1.0"
description = ""
authors = ["bgrobelsek <bruno.grobelsek@hotmail.com>"]
readme = "README.md"

package-mode = false

[tool.poetry.dependencies]
python = "^3.9"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-html = "^4.1.1"
pytest-xdist = "^3.6.1"
requests = "^2.32.3"
docker = "^7.1.0"
faker = "^30.0.0"
python-slugify = "^8.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
